ISSUE #751 moved all parser types into header
